+++ api/candidate.go
@@ -35,7 +35,7 @@ func GetCandidate(w http.ResponseWriter, r *http.Request) {
 		Result: struct {
 			Candidate Candidate `json:"candidate"`
 		}{
-			Candidate: makeResponseCandidate(*candidate, cState),
+			Candidate: makeResponseCandidate(*candidate),
 		},
 	})
 }