+++ core/transaction/sell_all_coin.go
@@ -141,6 +141,7 @@ func (data SellAllCoinData) Run(sender types.Address, tx *Transaction, context *
 		common.KVPair{Key: []byte("tx.coin_to_buy"), Value: []byte(data.CoinToBuy.String())},
 		common.KVPair{Key: []byte("tx.coin_to_sell"), Value: []byte(data.CoinToSell.String())},
 		common.KVPair{Key: []byte("tx.return"), Value: []byte(value.String())},
+		common.KVPair{Key: []byte("tx.sell_amount"), Value: []byte(amountToSell.String())},
 	}
 
 	return Response{